[BAD-1608] - Develop the cdh610 shim.
