+++ common/common-shim/src/test/java/org/pentaho/hadoop/shim/common/format/parquet/PentahoParquetInputFormatTest.java
@@ -34,10 +34,10 @@ import java.util.TimeZone;
 import java.util.TreeSet;
 
 import org.apache.hadoop.fs.Path;
-//#if shim_type=="HDP" || shim_type=="EMR" || shim_type=="HDI" || shim_name=="mapr60" || shim_name=="cdh601"
+//#if shim_type=="HDP" || shim_type=="EMR" || shim_type=="HDI" || shim_name=="mapr60" || shim_name=="cdh610"
 import org.apache.parquet.hadoop.ParquetInputSplit;
 //#endif
-//#if shim_type=="CDH" && shim_name!="cdh601" || shim_type=="MAPR" && shim_name!="mapr60"
+//#if shim_type=="CDH" && shim_name!="cdh610" || shim_type=="MAPR" && shim_name!="mapr60"
 //$import parquet.hadoop.ParquetInputSplit;
 //#endif
 import org.junit.Test;
