--- shims/cdh601/assemblies/cdh601-shim/src/assembly/assembly.xml
@@ -1,32 +1,0 @@
-<assembly>
-  <id>plugin</id>
-  <baseDirectory>cdh601</baseDirectory>
-  <formats>
-    <format>zip</format>
-  </formats>
-  <fileSets>
-    <fileSet>
-      <directory>${basedir}/src/main/resources</directory>
-      <outputDirectory>/</outputDirectory>
-    </fileSet>
-    <fileSet>
-      <directory>${oss.license.directory}</directory>
-      <outputDirectory>/</outputDirectory>
-    </fileSet>
-    <fileSet>
-      <directory>${project.build.directory}/cdh601</directory>
-      <outputDirectory>/</outputDirectory>
-    </fileSet>
-  </fileSets>
-  <dependencySets>
-    <dependencySet>
-      <outputDirectory>/</outputDirectory>
-      <useTransitiveDependencies>false</useTransitiveDependencies>
-      <useTransitiveFiltering>false</useTransitiveFiltering>
-      <includes>
-        <include>org.pentaho:pentaho-hadoop-shims-cdh601</include>
-        <include>org.pentaho:pentaho-hadoop-shims-cdh601-hbase-comparators</include>
-      </includes>
-    </dependencySet>
-  </dependencySets>
-</assembly>
