--- shims/cdh601/assemblies/cdh601-shim/src/main/resources/config.properties
@@ -1,19 +1,0 @@
-# Friendly name for this configuration
-name=Cloudera CDH 6.0.1
-
-# Comma-separated list of directories and files to make available to this
-# configuration. Any resources found here will overwrite ones in lib/.
-classpath=
-
-# Comma-separated list of paths that contain native libraries to load. These
-# could be added to LD_LIBRARY_PATH or set with -Djava.library.path instead.
-library.path=
-
-# Comma-separated list of classes or package names to explicitly ignore when
-# loading classes from the resources within this Hadoop configuration directory
-# or the classpath property
-# e.g.: org.apache.commons.log,org.apache.log4j
-# Note, the two packages above are automatically included for all configurations
-ignore.classes=org.apache.derby.iapi.services
-
-mr1.java.system.hadoop.cluster.path.separator=:
