--- shims/cdh601/client/src/assembly/assembly.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<assembly>
-  <id>plugin</id>
-  <baseDirectory>cdh601</baseDirectory>
-  <formats>
-    <format>zip</format>
-  </formats>
-  <dependencySets>
-    <dependencySet>
-      <outputDirectory>lib/client</outputDirectory>
-      <useTransitiveDependencies>true</useTransitiveDependencies>
-      <useTransitiveFiltering>true</useTransitiveFiltering>
-      <includes>
-        <include>org.xerial.snappy:snappy-java</include>
-        <include>org.apache.hadoop:hadoop-aws</include>
-        <include>org.apache.hadoop:hadoop-hdfs</include>
-        <include>org.apache.hadoop:hadoop-common</include>
-        <include>org.apache.hadoop:hadoop-auth</include>
-        <include>org.apache.hadoop:hadoop-mapreduce-client-core</include>
-        <include>org.apache.hadoop:hadoop-mapreduce-client-app</include>
-        <include>org.apache.hadoop:hadoop-mapreduce-client-common</include>
-        <include>org.apache.hadoop:hadoop-mapreduce-client-jobclient</include>
-        <include>org.apache.hadoop:hadoop-mapreduce-client-shuffle</include>
-        <include>org.apache.hadoop:hadoop-annotations</include>
-        <include>org.apache.hadoop:hadoop-yarn-api</include>
-        <include>org.apache.hadoop:hadoop-yarn-client</include>
-        <include>org.apache.hadoop:hadoop-yarn-common</include>
-        <include>org.apache.hadoop:hadoop-yarn-server-applicationhistoryservice</include>
-        <include>org.apache.hadoop:hadoop-yarn-server-common</include>
-        <include>org.apache.hadoop:hadoop-yarn-server-resourcemanager</include>
-        <include>org.apache.hadoop:hadoop-yarn-server-web-proxy</include>
-      </includes>
-      <excludes>
-        <exclude>junit:*</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-server-nodemanager</exclude>
-        <exclude>commons-collections:commons-collections</exclude>
-        <exclude>org.apache.zookeeper:zookeeper</exclude>
-        <exclude>com.google.guava:*</exclude>
-        <exclude>xml-apis:xml-apis</exclude>
-        <exclude>aws-java-sdk-bundle</exclude>
-        <exclude>*:tests:*</exclude>
-      </excludes>
-    </dependencySet>
-  </dependencySets>
-</assembly>
