--- shims/cdh601/default/src/assembly/assembly.xml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<assembly>
-  <id>plugin</id>
-  <baseDirectory>cdh601</baseDirectory>
-  <formats>
-    <format>zip</format>
-  </formats>
-  <dependencySets>
-    <dependencySet>
-      <outputDirectory>lib</outputDirectory>
-      <useTransitiveDependencies>true</useTransitiveDependencies>
-      <useTransitiveFiltering>false</useTransitiveFiltering>
-      <includes>
-        <include>org.xerial.snappy:snappy-java</include>
-        <include>pentaho:pentaho-hdfs-vfs</include>
-        <include>org.apache.hive:hive-jdbc</include>
-        <include>org.apache.oozie:oozie-core</include>
-        <include>org.apache.oozie:oozie-client</include>
-        <include>org.apache.sqoop:sqoop</include>
-        <include>org.apache.pig:pig</include>
-        <include>com.twitter:parquet-pig</include>
-        <include>com.twitter:parquet-hadoop-bundle</include>
-        <include>org.apache.hbase:hbase-protocol-shaded</include>
-        <include>com.codahale.metrics:metrics-core</include>
-        <include>org.apache.hbase.thirdparty:hbase-shaded-miscellaneous</include>
-        <include>org.apache.hbase.thirdparty:hbase-shaded-protobuf</include>
-        <include>org.apache.hbase.thirdparty:hbase-shaded-netty</include>
-        <include>org.python:jython</include>
-        <include>org.apache.hbase:hbase-hadoop2-compat</include>
-        <include>org.apache.hbase:hbase-metrics</include>
-        <include>org.apache.hbase:hbase-metrics-api</include>
-      </includes>
-      <excludes>
-        <exclude>log4j:log4j</exclude>
-        <exclude>org.mockito:mockito-all</exclude>
-        <exclude>junit:*</exclude>
-        <exclude>com.google.guava:*</exclude>
-        <exclude>xml-apis:xml-apis</exclude>
-        <exclude>commons-collections:commons-collections</exclude>
-        <exclude>org.apache.zookeeper:zookeeper</exclude>
-        <exclude>org.apache.hadoop:hadoop-annotations</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-api</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-client</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-common</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-server-applicationhistoryservice</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-server-common</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-server-resourcemanager</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-server-web-proxy</exclude>
-        <exclude>*:tests:*</exclude>
-        <exclude>*:hive-exec:*:core</exclude>
-        <exclude>org.apache.curator:apache-curator:pom</exclude>
-        <!--hive-exec excludes-->
-        <exclude>*:logredactor:*</exclude>
-        <exclude>*:datanucleus-core:*</exclude>
-        <exclude>*:hbase-common:*</exclude>
-        <exclude>*:ST4:*</exclude>
-      </excludes>
-    </dependencySet>
-    <dependencySet>
-      <outputDirectory>lib</outputDirectory>
-      <useTransitiveDependencies>true</useTransitiveDependencies>
-      <useTransitiveFiltering>true</useTransitiveFiltering>
-      <includes>
-        <include>org.apache.hive:hive-serde</include>
-        <include>org.apache.hive:hive-common</include>
-        <include>org.apache.hive:hive-exec</include>
-        <include>org.apache.hive:hive-service</include>
-        <include>org.apache.hive:hive-metastore</include>
-        <include>dk.brics.automaton:automaton</include>
-      </includes>
-      <excludes>
-        <exclude>log4j:log4j</exclude>
-        <exclude>org.mockito:mockito-all</exclude>
-        <exclude>junit:*</exclude>
-        <exclude>com.google.guava:*</exclude>
-        <exclude>xml-apis:xml-apis</exclude>
-        <exclude>commons-collections:commons-collections</exclude>
-        <exclude>org.apache.zookeeper:zookeeper</exclude>
-        <exclude>org.apache.hadoop:hadoop-annotations</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-api</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-client</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-common</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-server-applicationhistoryservice</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-server-common</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-server-resourcemanager</exclude>
-        <exclude>org.apache.hadoop:hadoop-yarn-server-web-proxy</exclude>
-        <exclude>*:tests:*</exclude>
-        <exclude>*:hive-exec:*:core</exclude>
-        <exclude>org.apache.curator:apache-curator:pom</exclude>
-        <!--hive-exec excludes-->
-        <exclude>*:logredactor:*</exclude>
-        <exclude>*:datanucleus-core:*</exclude>
-        <exclude>*:hbase-common:*</exclude>
-        <exclude>*:ST4:*</exclude>
-        <exclude>com.fasterxml.jackson.core:jackson-databind</exclude>
-      </excludes>
-    </dependencySet>
-  </dependencySets>
-</assembly>
