--- shims/cdh601/impl/src/main/java/org/pentaho/hadoop/shim/cdh514/authorization/ShimNoOpHadoopAuthorizationService.java
@@ -1,35 +1,0 @@
-/*******************************************************************************
-*
-* Pentaho Big Data
-*
-* Copyright (C) 2002-2019 by Hitachi Vantara : http://www.pentaho.com
-*
-*******************************************************************************
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with
-* the License. You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-******************************************************************************/
-
-package org.pentaho.hadoop.shim.cdh601.authorization;
-
-
-import org.pentaho.hadoop.shim.cdh601.HadoopShim;
-import org.pentaho.hadoop.shim.common.CommonHadoopShim;
-import org.pentaho.hadoop.shim.common.authorization.NoOpHadoopAuthorizationService;
-
-public class ShimNoOpHadoopAuthorizationService extends NoOpHadoopAuthorizationService {
-
-  @Override protected CommonHadoopShim getHadoopShim() {
-    return new HadoopShim();
-  }
-}
