--- shims/cdh601/impl/src/main/resources/META-INF/services/org.pentaho.hadoop.shim.common.authorization.NoOpHadoopAuthorizationService
@@ -1,1 +1,0 @@
-org.pentaho.hadoop.shim.cdh601.authorization.ShimNoOpHadoopAuthorizationService
