--- shims/cdh601/pom.xml
@@ -1,37 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.pentaho</groupId>
-    <artifactId>pentaho-hadoop-shims-list</artifactId>
-    <version>8.2.0.0-SNAPSHOT</version>
-  </parent>
-  <artifactId>pentaho-hadoop-shims-cdh601-reactor</artifactId>
-  <packaging>pom</packaging>
-  <properties>
-    <shim.name>cdh601</shim.name>
-    <!-- default folder -->
-	<org.xerial.snappy.version>1.1.1.3</org.xerial.snappy.version>
-    <org.apache.hive.version>2.1.1-cdh6.0.1</org.apache.hive.version>
-    <pig.version>0.17.0-cdh6.0.1</pig.version>
-    <parquet-pig.version>1.9.0-cdh6.0.1</parquet-pig.version>
-    <sqoop.version>1.4.7-cdh6.0.1</sqoop.version>
-    <org.apache.oozie.version>5.0.0-cdh6.0.1</org.apache.oozie.version>
-    <!-- client folder -->
-    <org.apache.hadoop.version>3.0.0-cdh6.0.1</org.apache.hadoop.version>
-    <!-- pmr folder -->
-    <org.apache.hbase.version>2.0.0-cdh6.0.1</org.apache.hbase.version>
-    <zookeeper.version>3.4.5-cdh6.0.1</zookeeper.version>
-    <com.codahale.metrics.version>3.0.1</com.codahale.metrics.version>
-    <slf4j-api.version>1.7.7</slf4j-api.version>
-    <slf4j-log4j12.version>1.7.7</slf4j-log4j12.version>
-  </properties>
-  <modules>
-    <module>client</module>
-    <module>default</module>
-    <module>pmr</module>
-    <module>impl</module>
-    <module>hbase-comparators</module>
-    <module>assemblies</module>
-  </modules>
-</project>
