+++ shims/cdh610/assemblies/cdh610-shim/src/assembly/assembly.xml
@@ -1,0 +1,32 @@
+<assembly>
+  <id>plugin</id>
+  <baseDirectory>cdh610</baseDirectory>
+  <formats>
+    <format>zip</format>
+  </formats>
+  <fileSets>
+    <fileSet>
+      <directory>${basedir}/src/main/resources</directory>
+      <outputDirectory>/</outputDirectory>
+    </fileSet>
+    <fileSet>
+      <directory>${oss.license.directory}</directory>
+      <outputDirectory>/</outputDirectory>
+    </fileSet>
+    <fileSet>
+      <directory>${project.build.directory}/cdh610</directory>
+      <outputDirectory>/</outputDirectory>
+    </fileSet>
+  </fileSets>
+  <dependencySets>
+    <dependencySet>
+      <outputDirectory>/</outputDirectory>
+      <useTransitiveDependencies>false</useTransitiveDependencies>
+      <useTransitiveFiltering>false</useTransitiveFiltering>
+      <includes>
+        <include>org.pentaho:pentaho-hadoop-shims-cdh610</include>
+        <include>org.pentaho:pentaho-hadoop-shims-cdh610-hbase-comparators</include>
+      </includes>
+    </dependencySet>
+  </dependencySets>
+</assembly>
