+++ shims/cdh610/assemblies/cdh610-shim/src/main/resources/hdfs-site.xml
