+++ shims/cdh610/impl/src/main/java/org/apache/hadoop/hive/jdbc/HiveDriver.java
