+++ shims/pom.xml
@@ -382,7 +382,7 @@
         <module>cdh512</module>
         <module>cdh513</module>
         <module>cdh514</module>
-        <module>cdh601</module>
+        <module>cdh610</module>
         <module>mapr520</module>
         <module>mapr60</module>
         <module>emr59</module>
@@ -460,6 +460,18 @@
       </activation>
       <modules>
         <module>cdh514</module>
+      </modules>
+    </profile>
+
+    <profile>
+      <id>cdh610</id>
+      <activation>
+        <property>
+          <name>!skipDefault</name>
+        </property>
+      </activation>
+      <modules>
+        <module>cdh610</module>
       </modules>
     </profile>
